Backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
langchain
google-genai
scikit-learn
faiss-cpu
